refactored codebase by changing About.tsx and Home.txt to Pages.tsx
