+++ frontend/src/App.tsx
@@ -2,15 +2,15 @@ import { MantineProvider } from "@mantine/core";
 import { theme } from "./theme";
 import "./index.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Home from "./pages/Home/Home";
-import About from "./pages/About/About";
-import Ryder from "./pages/RyderModal/Home";
-import Michelle from "./pages/MHome/Home";
-import Rules from "./pages/Rules/Home";
-import Lobby from "./pages/Lobby/About";
+import Home from "./pages/Home/Pages";
+import About from "./pages/About/Pages";
+import Ryder from "./pages/RyderModal/Pages";
+import Michelle from "./pages/MHome/Pages";
+import Rules from "./pages/Rules/Pages";
+import Lobby from "./pages/Lobby/Pages";
 import Night from "./pages/Night/Night";
-import Kitchen from "./pages/Kitchen/About";
-import Day from "./pages/DayText/About";
+import Kitchen from "./pages/Kitchen/Pages";
+import Day from "./pages/DayText/Pages";
 
 export default function App() {
   return (
