+++ frontend/src/pages/DayText/Pages.tsx
