+++ frontend/src/pages/Rules/Pages.tsx
